extract resolvers into separate files
