+++ app.js
@@ -1,15 +1,8 @@
+const bodyParser = require('body-parser')
 const express = require('express')
 const schema = require('./schema')
-const resolve = require('./resolvers')
+const root = require('./resolvers')
 const { graphqlHTTP } = require('express-graphql')
-const bodyParser = require('body-parser')
-
-const root = {
-  products: resolve.getProducts,
-  addProduct: resolve.addProduct,
-  editProductBalance: resolve.editProductBalance,
-  deleteProducts: resolve.deleteProducts
-}
 
 const app = express()
