+++ resolvers/getProducts.js
@@ -1,0 +1,5 @@
+const models = require("../models");
+
+const getProducts = async () => await models.Product.findAll()
+
+module.exports = getProducts;
